Lightwave 3D Exporter: -Added some Status Bar messages.
DevWork:
-Built & separated Visual Studio 2008 & 2010 versions of WMV's project files.
-StormLib is now included in the VS Project files and will be compiled prior to WMV being built.
-Recompiled & renamed all the .lib files.
-Removed 2 unnecessary lib files from our build of StormLib.
-Minor code fixes.

Wiki:
-Updated Windows Compilation page.

+++ trunk/src/modelexport_ms3d.cpp
@@ -245,7 +245,7 @@ void ExportMS3D_M2(Attachment *att, Model *m, const char *fn, bool init)
 			joint.rotation[1] = 0;
 			joint.rotation[2] = 0;
 
-			Vec3D p = FixPivot(m, i, m->bones[i].pivot);
+			Vec3D p = FixPivot(m, (int)i, m->bones[i].pivot);
 			joint.position[0] = p.x;
 			joint.position[1] = p.y;
 			joint.position[2] = p.z;
